backend/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# 时序数据库
influxdb-client==1.38.0

# 缓存
redis==5.0.1

# MQTT
paho-mqtt==1.6.1

# 数据验证
pydantic==2.5.0
pydantic-settings==2.1.0

# 工具库
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# 其他
requests==2.31.0
aiofiles==23.2.1